% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{linreg}
\alias{linreg}
\title{A S3 class provide linear regression  that is similar as lm().}
\usage{
linreg(formula, data, qr = FALSE)
}
\arguments{
\item{formula}{Formula}

\item{data}{Dataframe}

\item{qr}{Boolean}
}
\value{
linreg object(list) contain all the calculated values regarding the linear regression
}
\description{
Author: Qinyuqn Qi
}
