% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{myclass}
\alias{myclass}
\alias{linreg}
\title{My S3 Class}
\usage{
linreg(formula, data, qr = FALSE)
}
\arguments{
\item{formula}{Formula}

\item{data}{Dataframe}

\item{qr}{Boolean}
}
\value{
linreg object(list) contain all the calculated values regarding the linear regression
}
\description{
This is a simple S3 class for demonstration purposes.
}
\details{
A S3 class provide linear regression  that is similar as lm().

Author: Qinyuqn Qi
}
\section{Fields}{

\describe{
\item{\code{data}}{A numeric vector.}
}}

